requirements.txt

chainlit>=1.1.0
openai>=1.30.0
pyyaml>=6.0.1
tiktoken>=0.7.0
numpy>=1.26.0
